Disable automatic jekyll build of GitHub page
The way this jekyll template is built is with yarn which is a
wrapper which builds the static jekyll site. As such, there is
no need for GitHub pages to attempt a native jekyll build as
this will not work at all.

Instead, the `npm run deploy` will take the generated files and
commit them to the repo's `gh-pages` branch for publication. This
has to be done from the CLI (or through a CI/CD system).